Trekk ut kode i egne komponenter
