+++ src/overgangsstønad/steg/8-dokumentasjon/ManglendeInformasjonAlert.tsx
@@ -23,7 +23,7 @@ const ManglendeInformasjonAlert: FC = () => (
         }}
       >
         <LocaleTekst tekst="dokumentasjon.alert.link.fylleInn" />
-      </Link>{' '}
+      </Link>
       <LocaleTekst tekst="dokumentasjon.alert.manglende" />
     </AlertStripe>
   </KomponentGruppe>
