+++ src/overgangsstønad/steg/8-dokumentasjon/SendSøknad.tsx
@@ -1,21 +1,16 @@
 import React, { FC } from 'react';
-import KnappBase from 'nav-frontend-knapper';
-import LocaleTekst from '../../../language/LocaleTekst';
 import { IStatus } from '../../../arbeidssøkerskjema/innsending/typer';
 import { ISøknad, LocationStateSøknad } from '../../../models/søknad/søknad';
 import { parseISO } from 'date-fns';
 import { useSøknad } from '../../../context/SøknadContext';
 import { useHistory, useLocation } from 'react-router';
-import { Normaltekst } from 'nav-frontend-typografi';
 import { RoutesOvergangsstonad } from '../../routing/routesOvergangsstonad';
-import SeksjonGruppe from '../../../components/gruppe/SeksjonGruppe';
-import { StyledKnapper } from '../../../arbeidssøkerskjema/komponenter/StyledKnapper';
 import {
   mapBarnTilEntenIdentEllerFødselsdato,
   mapBarnUtenBarnepass,
   sendInnSøknad,
 } from '../../../innsending/api';
-import { hentForrigeRoute, hentNesteRoute } from '../../../utils/routing';
+import { hentNesteRoute } from '../../../utils/routing';
 import { unikeDokumentasjonsbehov } from '../../../utils/søknad';
 import { useSpråkContext } from '../../../context/SpråkContext';
 import { useIntl } from 'react-intl';
@@ -27,16 +22,14 @@ import { logDokumetasjonsbehov } from '../../../utils/amplitude';
 import { ESkjemanavn } from '../../../utils/skjemanavn';
 import InnsendingFeilet from './InnsendingFeilet';
 import ManglendeInformasjonAlert from './ManglendeInformasjonAlert';
+import TilbakeSendAvBrytKnapper from './TilbakeSendAvBrytKnapper';
+import SaksbehandleINyLøsningKnapp from './SaksbehandleINyLøsning';
 
 export interface Innsending {
   status: string;
   melding: string;
   venter: boolean;
 }
-
-const validerSøkerBosattINorgeSisteTreÅr = (søknad: ISøknad) => {
-  return søknad.medlemskap.søkerBosattINorgeSisteTreÅr;
-};
 
 const SendSøknadKnapper: FC = () => {
   const { søknad, settSøknad } = useSøknad();
@@ -46,16 +39,16 @@ const SendSøknadKnapper: FC = () => {
   const history = useHistory();
   const nesteRoute = hentNesteRoute(RoutesOvergangsstonad, location.pathname);
   const intl = useIntl();
-  const forrigeRoute = hentForrigeRoute(
-    RoutesOvergangsstonad,
-    location.pathname
-  );
 
   const [innsendingState, settinnsendingState] = React.useState<Innsending>({
     status: IStatus.KLAR_TIL_INNSENDING,
     melding: `Søknad kan sendes`,
     venter: false,
   });
+
+  const validerSøkerBosattINorgeSisteTreÅr = (søknad: ISøknad) => {
+    return søknad.medlemskap.søkerBosattINorgeSisteTreÅr;
+  };
 
   const oppdaterBarnLabels = (barn: IBarn[]) => {
     const oppdaterteBarn = barn.map((barnet: any) => {
@@ -86,7 +79,7 @@ const SendSøknadKnapper: FC = () => {
     return oppdaterteBarn;
   };
 
-  const sendSøknad = (søknad: ISøknad) => {
+  const sendSøknad = (søknad: ISøknad): void => {
     const barnMedEntenIdentEllerFødselsdato = mapBarnUtenBarnepass(
       mapBarnTilEntenIdentEllerFødselsdato(søknad.person.barn)
     );
@@ -141,54 +134,17 @@ const SendSøknadKnapper: FC = () => {
         <ManglendeInformasjonAlert />
       )}
       {toggles[ToggleName.visSkalBehandlesINySaksbehandling] && (
-        <>
-          <div className={'sentrert'}>
-            <KnappBase
-              type={'hoved'}
-              onClick={() =>
-                !innsendingState.venter &&
-                sendSøknad({ ...søknad, skalBehandlesINySaksbehandling: true })
-              }
-              spinner={innsendingState.venter}
-            >
-              Saksbehandle i ny løsning
-            </KnappBase>
-          </div>
-          <Normaltekst style={{ marginBottom: 30 }}>
-            (For å IKKE forsøke automatisk journalføring og blankettbehandling
-            trykk på "Saksbehandle i ny løsning" knapp)
-          </Normaltekst>
-        </>
+        <SaksbehandleINyLøsningKnapp
+          innsendingState={innsendingState}
+          søknad={søknad}
+          sendSøknad={sendSøknad}
+        />
       )}
-      <SeksjonGruppe className={'sentrert'}>
-        <StyledKnapper>
-          <KnappBase
-            className={'tilbake'}
-            type={'standard'}
-            onClick={() => history.push(forrigeRoute.path)}
-          >
-            <LocaleTekst tekst={'knapp.tilbake'} />
-          </KnappBase>
-
-          {validerSøkerBosattINorgeSisteTreÅr(søknad) && (
-            <KnappBase
-              type={'hoved'}
-              onClick={() => !innsendingState.venter && sendSøknad(søknad)}
-              className={'neste'}
-              spinner={innsendingState.venter}
-            >
-              <LocaleTekst tekst={'knapp.sendSøknad'} />
-            </KnappBase>
-          )}
-          <KnappBase
-            className={'avbryt'}
-            type={'flat'}
-            onClick={() => history.push(RoutesOvergangsstonad[0].path)}
-          >
-            <LocaleTekst tekst={'knapp.avbryt'} />
-          </KnappBase>
-        </StyledKnapper>
-      </SeksjonGruppe>
+      <TilbakeSendAvBrytKnapper
+        innsendingState={innsendingState}
+        søknad={søknad}
+        sendSøknad={sendSøknad}
+      />
     </>
   );
 };
